requirements.txt
appdirs==1.4.4
asgiref==3.7.2
beautifulsoup4==4.12.3
certifi==2023.7.22
Django==4.2.7
django-crispy-forms
crispy_bootstrap5
django-environ==0.11.2
django-storages==1.14.2
django-widget-tweaks==1.5.0
email_validator==2.2.0
google-api-core==2.14.0
google-auth==2.23.4
google-cloud-appengine-logging==1.4.5
google-cloud-audit-log==0.3.0
google-cloud-core==2.3.3
google-cloud-logging==3.11.3
google-cloud-secret-manager==2.16.4
google-cloud-storage==2.13.0
google-crc32c==1.5.0
gunicorn==21.2.0
html5lib==1.1
httpx==0.27.0
lxml==5.1.0
numpy==1.26.3
oauthlib==3.2.2
openai==1.13.3
pandas==2.2.0
pillow==10.2.0
psycopg2-binary
python-dateutil==2.8.2
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
soupsieve==2.5
yfinance==0.2.36
google-api-python-client
social-auth-app-django
google-auth-httplib2
google-auth-oauthlib
msal
google-cloud-tasks==2.16.0  # Latest stable version as of now
